Debug opt, test coverage, and a better embed conditional (#7)
* Move async/await out of visitor function

Fix #5

* Add debug: true option, turn off output otherwise

* Only embed when twitter link on a line by itself

Otherwise it does really wierd things to inline links to tweets, which
looks pretty bad.

* Add tests

With 100% coverage, just for bonus points.

TODO: Still need tests for the gatsby-browser.js file, not sure the best
way to go about that.  Some kind of browser shim thing?

* Add package-lock.json file

* Ignore coverage folder

* Make blockquote string not as network-dependent

I just realized that when/if I change my twitter handle, it'll break
tests for this module.  Oops!  Fix that by trimming the string to just
the beginning part.